slim down Lobby/index
